# flight-plugin/openapi.yaml
openapi: 3.0.1
info:
  title: Flights Search API
  description: Finds up to date and detailed flight tickets and get your calendar events
  version: 0.1.0
servers:
  - url: https://2t7r3lvw-8080.usw2.devtunnels.ms/
paths:
  /get-accounts:
    get:
      operationId: getAccounts
      summary: Get your account info
      description: Returns your firstname, lastname, birthday, phonenumber, id and email
      parameters:
        - name: firstName
          in: query
          description: firstName
          required: false
          schema:
            type: string
        - name: lastName
          in: query
          description: lastName
          required: false
          schema:
            type: string
        - name: phoneNumber
          in: query
          description: phoneNumber
          required: false
          schema:
            type: string
        - name: id
          in: query
          description: id
          required: false
          schema:
            type: string
        - name: email
          in: query
          description: email
          required: false
          schema:
            type: string
        - name: birthday
          in: query
          description: birthday
          required: false
          schema:
            type: string
        - name: payment
          in: query
          description: payment method
          required: false
          schema: 
            $ref: '#/components/schemas/Payment'
      responses:
        '200':
          description: OK - Returns a list of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
  /get-flights:
    get:
      operationId: getFlights
      summary: Get a list of flights matching the specified criteria
      description: Returns a list of flights, optionally filtered by providing the departure airport, destination airport, departure date range start and end, and connections as query parameters
      parameters:
        - name: departure
          in: query
          description: The departure airport to search in
          required: true
          schema:
            type: string
        - name: destination
          in: query
          description: The destination airport to search in
          required: true
          schema:
            type: string
        - name: date
          in: query
          description: The exact departure date of the flight
          required: false
          schema:
            type: string
        - name: dateRangeStart
          in: query
          description: The earliest possible departure date of the flight
          required: false
          schema:
            type: string
        - name: dateRangeEnd
          in: query
          description: The latest possible departure date of the flight
          required: false
          schema:
            type: string
        - name: connections
          in: query
          description: The number of connections
          required: false
          schema:
            type: integer
        - name: class
          in: query
          description: The class of the flight
          required: false
          schema:
            $ref: '#/components/schemas/FlightClassesEnum'
      responses:
        '200':
          description: OK - Returns a list of flights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Flight'
  /book:
    get:
      operationId: bookFlights
      summary: Book flight
      description: book flight beased on search on copilot
      parameters:
        - name: airline
          in: query
          description: airline name
          required: true
          schema:
            type: string
        - name: departure
          in: query
          description: The departure airport to search in
          required: true
          schema:
            type: string
        - name: destination
          in: query
          description: The destination airport to search in
          required: true
          schema:
            type: string
        - name: date
          in: query
          description: The exact departure date of the flight
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            text/plain:
              schema:
                type: string
  /get-calendar:
    get:
      operationId: getCalendar
      summary: Get you calendar events
      description: Returns your calendar events, optionally filtered by providing the event name, start time, and end time as query parameters
      parameters:
        - name: event
          in: query
          description: The event name
          required: false
          schema:
            type: string
        - name: meetingStart
          in: query
          description: The start time of the event
          required: false
          schema:
            type: string
        - name: meetingEnd
          in: query
          description: The end time of the event
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK - Returns a list of calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
components:
  schemas:
    FlightClassesEnum:
      type: string
      enum:
        - Economy
        - Business
        - First
    Payment:
      type: object
      properties:
        number:
          type: string
          description: number
        expirationDate:
          type: string
          description: expirationDate
        cvv:
          type: string
          description: cvv
        name:
          type: string
          description: name
    Flight:
      type: object
      properties:
        airline:
          type: string
          description: The airline operating the flight
        flight_number:
          type: string
          description: The flight number
        departure_airport:
          type: string
          description: The departure airport of the flight
        destination_airport:
          type: string
          description: The destination airport of the flight
        departure_time:
          type: string
          description: The depature time of the flight
        arrival_time:
          type: string
          description: The arrival time of the flight
        connections:
          type: integer
          description: The number of connections
        class:
          type: string
          description: The class of the flight ticket
          items:
            $ref: '#/components/schemas/FlightClassesEnum'
        price:
          type: integer
          description: The price of the ticket
        duration:
          type: string
          description: The duration of the flight
        ticket_link:
          type: string
          description: The link to the ticket
    Event:
      type: object
      properties:
        event:
          type: string
          description: The name of the event
        meetingStart:
          type: string
          description: The start time of the event
        meetingEnd:
          type: string
          description: The end time of the event
    Account:
      type: object
      properties:
        firstName:
          type: string
          description: firstName
        lastName:
          type: string
          description: lastName
        id:
          type: string
          description: id
        phoneNumber:
          type: string
          description: phoneNumber
        email:
          type: string
          description: email
        birthday:
          type: string
          description: birthday